requirements.txt
# 基础依赖
xtquant>=1.0.0
python-yaml>=6.0.0
pandas>=1.3.0
numpy>=1.21.0

# 数据分析和可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 工具库
loguru>=0.6.0  # 日志工具
click>=8.0.0  # 命令行工具